# ai-platform-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI平台智能分析系统 API",
    "description": "AI平台智能分析系统 - 用于分析客服对话并检测规避责任行为",
    "version": "1.0.0",
    "contact": {
      "name": "AI平台开发团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "本地开发环境"
    },
    {
      "url": "http://your-production-server/api/v1",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "使用Bearer Token进行身份验证，格式：Bearer YOUR_API_KEY"
      }
    },
    "schemas": {
      "AnalysisRequest": {
        "type": "object",
        "properties": {
          "work_id": {
            "type": "integer",
            "description": "指定工单ID",
            "example": 1234567890,
            "nullable": true
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "description": "开始时间",
            "nullable": true
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "description": "结束时间",
            "nullable": true
          },
          "hours_back": {
            "type": "integer",
            "minimum": 1,
            "maximum": 168,
            "default": 24,
            "description": "向前查找小时数"
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000,
            "default": 100,
            "description": "处理限制数量"
          }
        }
      },
      "HealthStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["healthy", "unhealthy"],
            "description": "健康状态"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "检查时间戳"
          },
          "components": {
            "type": "object",
            "properties": {
              "database": {
                "type": "string",
                "description": "数据库状态"
              },
              "scheduler": {
                "type": "string",
                "description": "调度器状态"
              }
            }
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据",
            "nullable": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "错误详情"
          }
        }
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "tags": ["基础信息"],
        "summary": "获取应用基础信息",
        "description": "获取应用基础信息，无需认证",
        "security": [],
        "responses": {
          "200": {
            "description": "成功返回应用信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "app_name": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "environment": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "docs_url": {
                      "type": "string"
                    },
                    "api_prefix": {
                      "type": "string"
                    },
                    "auth_required": {
                      "type": "boolean"
                    },
                    "auth_header": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["基础信息"],
        "summary": "简单健康检查",
        "description": "简单的健康检查接口，无需认证",
        "security": [],
        "responses": {
          "200": {
            "description": "健康状态正常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "app_name": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "environment": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analysis/manual/batch": {
      "post": {
        "tags": ["分析管理"],
        "summary": "手动触发批量分析",
        "description": "手动触发批量分析任务",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "批量分析任务执行成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "result": {
                          "type": "object",
                          "description": "分析结果"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/manual/work-id": {
      "post": {
        "tags": ["分析管理"],
        "summary": "按工单ID执行分析",
        "description": "针对指定工单ID执行分析任务",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AnalysisRequest"
                  },
                  {
                    "type": "object",
                    "required": ["work_id"]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "工单分析执行成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "result": {
                          "type": "object",
                          "description": "分析结果"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "工单未找到或无回复记录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/manual/time-range": {
      "post": {
        "tags": ["分析管理"],
        "summary": "按时间段执行分析",
        "description": "针对指定时间段内的数据执行分析任务",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "时间段分析执行成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "time_range": {
                          "type": "object",
                          "properties": {
                            "start_time": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "end_time": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        },
                        "result": {
                          "type": "object",
                          "description": "分析结果"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/health": {
      "get": {
        "tags": ["系统管理"],
        "summary": "系统健康检查",
        "description": "详细的系统健康检查，包括数据库、调度器等组件状态",
        "responses": {
          "200": {
            "description": "系统健康",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            }
          },
          "503": {
            "description": "系统不健康",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            }
          }
        }
      }
    },
    "/system/scheduler/status": {
      "get": {
        "tags": ["调度器管理"],
        "summary": "获取调度器状态",
        "description": "获取调度器运行状态和任务信息",
        "responses": {
          "200": {
            "description": "调度器状态获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "scheduler_status": {
                          "type": "object",
                          "description": "调度器状态详情"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "获取调度器状态失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/scheduler/start": {
      "post": {
        "tags": ["调度器管理"],
        "summary": "启动调度器",
        "description": "启动定时任务调度器",
        "responses": {
          "200": {
            "description": "调度器启动成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "启动调度器失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/scheduler/stop": {
      "post": {
        "tags": ["调度器管理"],
        "summary": "停止调度器",
        "description": "停止定时任务调度器",
        "responses": {
          "200": {
            "description": "调度器停止成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "停止调度器失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "基础信息",
      "description": "应用基础信息和健康检查接口"
    },
    {
      "name": "分析管理",
      "description": "分析任务管理接口"
    },
    {
      "name": "系统管理", 
      "description": "系统健康检查接口"
    },
    {
      "name": "调度器管理",
      "description": "定时任务调度器管理接口"
    }
  ]
}